# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "McAfee SIEM Alarm Editor API",
    "description": "REST API for managing SIEM customers, rules, and alarms",
    "version": "2.0.0",
    "contact": {
      "name": "McAfee SIEM Alarm Editor",
      "email": "support@example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/customers/{id}/alarms": {
      "get": {
        "tags": ["Alarms"],
        "summary": "Get customer alarms",
        "description": "Retrieve all alarms for a specific customer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "Customer ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of alarms",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "alarms": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Alarm" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Alarms"],
        "summary": "Create alarm",
        "description": "Create a new alarm for a customer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "Customer ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "High Severity Alert" },
                  "severity": { "type": "integer", "minimum": 0, "maximum": 100 },
                  "match_value": { "type": "string", "example": "47|6000114" },
                  "match_field": { "type": "string", "default": "DSIDSigID" },
                  "condition_type": { "type": "integer", "default": 14 },
                  "note": { "type": "string" }
                },
                "required": ["name", "severity", "match_value"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Alarm created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "alarm": { "$ref": "#/components/schemas/Alarm" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Alarm": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "customer_id": { "type": "integer" },
          "name": { "type": "string", "example": "High Severity Alert" },
          "min_version": { "type": "string", "example": "11.6.14" },
          "severity": { "type": "integer", "minimum": 0, "maximum": 100 },
          "match_field": { "type": "string", "example": "DSIDSigID" },
          "match_value": { "type": "string", "example": "47|6000114" },
          "condition_type": { "type": "integer" },
          "assignee_id": { "type": "integer" },
          "esc_assignee_id": { "type": "integer" },
          "note": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
...
}